refactor: make logging configuration more consistent
* Use conserved set of logger names.
* Use `from` imports.
* Move around a few modules.
