+++ tests/test_executor.py
@@ -3,7 +3,7 @@ from asyncio import CancelledError
 import pytest
 from pytest_mock import MockerFixture
 from virtool_workflow import Workflow, hooks
-from virtool_workflow._executor import execute
+from virtool_workflow.runtime.executor import execute
 from virtool_workflow.events import Events
 from virtool_workflow.execution.states import COMPLETE, ERROR, RUNNING, WAITING
