+++ virtool_workflow/__init__.py
@@ -2,7 +2,7 @@
 from pyfixtures import fixture
 from virtool_workflow.decorator_api import step
 from virtool_workflow.workflow import Workflow
-from virtool_workflow._executor import execute
+from virtool_workflow.runtime.executor import execute
 from virtool_workflow._steps import WorkflowStep
 
 __all__ = [
