+++ virtool_workflow/api/samples.py
@@ -14,7 +14,7 @@ from virtool_workflow.api.utils import read_file_from_response, upload_file_via_
 from virtool_workflow.data_model.files import VirtoolFile, VirtoolFileFormat
 from virtool_workflow.data_model.samples import WFSample
 
-logger = getLogger("samples")
+logger = getLogger("api")
 
 
 class SampleProvider:
