+++ virtool_workflow/execution/hooks.py
@@ -1,11 +1,12 @@
-import logging
 from asyncio import gather
+from logging import getLogger
 from typing import Any, Callable, List
 
 from pyfixtures import FixtureScope
+
 from virtool_workflow.utils import coerce_to_coroutine_function
 
-logger = logging.getLogger(__name__)
+logger = getLogger("hooks")
 
 
 class Hook:
